Added functions to save and load HMM objects, and improved graphics

+++ glhmm/glhmm.py
@@ -40,7 +40,7 @@ class glhmm():
         Model for the mean. If `no` the mean of the timeseries will not be used to drive the states.
     model_beta : str, {'state', 'no'}, default 'state'
         Model for the beta. If `no` the regression coefficients will not be used to drive the states.
-    dirichlet_diag : int, default=10
+    dirichlet_diag : float, default=10
         The value of the diagonal of the Dirichlet distribution for the transition probabilities. 
         The higher the value, the more persistent the states will be. 
         Note that this value is relative; the prior competes with the data, so if the timeseries is very long,
